Replace expectations with sample repository
I've been trying to understand the underlying issues with the
expectation implementation but I was not able to fix the three different
type of race conditions I was able to find.

I've decided to take a different approach and see if we can track the
samples with a simple map series and pairs of samples.